Upload HW4 Autoencoders
Autoencoders, a group of architectures used for encoding compact representations of model inputs and then reconstructing them. This has a variety of real-world use
cases such as compression, pre-training encoder modules, and more. It is also closely related to the Variational Autoencoder models that we see later and which can be used to generate new synthetic data.

We implement a vanilla and then a stacked autoencoder model then train
each on Heart Failure Prediction and compare the results.